requirements.txt
aiohttp==3.9.5
json_stream==2.3.2
openai==1.31.0
python-dotenv==1.0.1
streamlit==1.35.0
tenacity==8.3.0
tiktoken==0.7.0
